+++ test/main.test.ts
@@ -10,6 +10,7 @@ test("login", async () => {
   await c.login()
   expect(c.csrfToken).not.toEqual("fetch")
 })
+
 test("logout", async () => {
   const c = createHttp()
   expect(c).toBeDefined()
@@ -17,6 +18,7 @@ test("logout", async () => {
   expect(c.csrfToken).not.toEqual("fetch")
   await c.logout()
 })
+
 test("badToken", async () => {
   const c = createHttp("DE")
   await c.login()